+++ feature/tasklist/api/src/main/java/ru/kpfu/itis/gureva/feature/tasklist/api/usecase/GetGroupNameUseCase.kt
@@ -1,5 +1,5 @@
 package ru.kpfu.itis.gureva.feature.tasklist.api.usecase
 
 interface GetGroupNameUseCase {
-    suspend operator fun invoke(id: Int): String?
+    suspend operator fun invoke(id: Long): String?
 }
