+++ feature/tasklist/api/src/main/java/ru/kpfu/itis/gureva/feature/tasklist/api/usecase/GetTasksUseCase.kt
@@ -3,5 +3,5 @@ package ru.kpfu.itis.gureva.feature.tasklist.api.usecase
 import ru.kpfu.itis.gureva.feature.tasklist.api.model.Task
 
 interface GetTasksUseCase {
-    suspend operator fun invoke(groupId: Int): List<Task>
+    suspend operator fun invoke(groupId: Long): List<Task>
 }
